requirement.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]~=1.2
livekit-agents[google]~=1.2
livekit-plugins-noise-cancellation~=0.2
python-dotenv